Dump of Lazarus Form Recovery Chrome extension
License of this work is UNKNOWN. I am creating a repo of it for:
a) learning & educational purposes.
  1. I personally wish to know how it works, how it does it's thing
  2. Chrome Web Store tells me this is extension violates it's policies
     and I would like to glean some sense of why
b) preservation in case it is ever deleted outright from the Chrome Web
Store. presently this extension is marked as violating their store
policies, and I can not find it on their web interface, however Chrome
will let me "enable" it in my browsers. I fret that someday I may have a
new install & be unable to bring it up.

Lazarus: Form Recovery was licensed as Freeware; an ambiguous term but
one certainly not confusable with open-source. This code is thus, to my
knowledge, available only for archival & educational purposes, and to
continue the Freeware license as able. I have no license or rights to
this code beyond fair use terms described above, and neither do you,
best I know. Godspeed.